+++ CHANGELOG.md
@@ -13,3 +13,13 @@ All notable changes to the react-use-localstate will be documented in this file.
 
 ### Added
 - Cross Tab Sync via Storage event listener
+
+## [1.2.0] - 2026-06-16
+
+## Fixed
+- Fixed Unsafe Double Type Casting
+
+## Added
+- Implemented Consistent Error Handling
+- Added Clear/Remove Function
+- Added Quota Limit Detection